requirements.txt
Django==5.1.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
geopy==2.4.1
stripe==7.11.0
django-rest-knox==4.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
daphne==4.0.0
djangorestframework-simplejwt==5.3.0
stripe==7.11.0
Pillow
python-dotenv==1.0.1
